rename http_server back to api_server
