+++ rust/sglang-server/src/api_server/common.rs
