+++ rust/sglang-server/src/api_server/openai/reasoning.rs
