+++ rust/sglang-server/src/api_server/openai/template.rs
