+++ rust/sglang-server/src/api_server/openai/tools.rs
