+++ rust/sglang-server/src/api_server/prefetch.rs
